+++ src/main/kotlin/aoc2019/SensorBoost.kt
@@ -1,8 +1,6 @@
 package aoc2019
 
-import java.math.BigInteger
-
-fun sensorBoost(param: Int, input: String): BigInteger {
+fun sensorBoost(param: Long, input: String): Long {
     val machine = IntCodeMachine(input)
     machine.sendInput(param)
     machine.run()
